Fix: move Chatboostadded to the correct file
