+++ docs/source/telegram.chatboostadded.rst
@@ -1,6 +1,8 @@
 ChatBoostAdded
 ==============
 
+.. versionadded:: NEXT.VERSION
+
 .. autoclass:: telegram.ChatBoostAdded
     :members:
     :show-inheritance:
